Refactoring...

+++ src/services/book/DeleteBookService.js
@@ -3,10 +3,15 @@ module.exports = class DeleteBookService {
   constructor(db) {this.db = db;}
 
   async execute(userId, bookId) {
-    // TODO: Checar se o userId existe.
+    // Checa se o usuário existe
+    const user = (await this.db.getUserById(userId));
+    if(!user) throw new Error("User does not exist.");
 
-    const book = await this.db.removeUserBook(userId, bookId);
+    // Checa se o livro existe
+    const book = (await this.db.getBookById(userId, bookId));
+    if(!book) throw new Error("Book does not exist.");
 
-    return book;
+    const deletedBook = await this.db.removeUserBook(userId, bookId);
+    return deletedBook;
   }
 }
